chore: move functions out of plugin into utils
